+++ client/lib/geometry.cpp
@@ -50,8 +50,8 @@ class Device {
     Eigen::Matrix<double, 3, NUM_TRANS_IN_UNIT> local_trans_positions;
 
     auto index = 0;
-    for (int y = 0; y < NUM_TRANS_Y; y++)
-      for (int x = 0; x < NUM_TRANS_X; x++)
+    for (size_t y = 0; y < NUM_TRANS_Y; y++)
+      for (size_t x = 0; x < NUM_TRANS_X; x++)
         if (!IS_MISSING_TRANSDUCER(x, y)) local_trans_positions.col(index++) = Eigen::Vector3d(x * TRANS_SIZE_MM, y * TRANS_SIZE_MM, 0);
 
     device->global_trans_positions = transform_matrix * local_trans_positions;
@@ -77,29 +77,29 @@ class AUTDGeometry : public Geometry {
  public:
   ~AUTDGeometry() override {}
 
-  int AddDevice(Vector3 position, Vector3 euler_angles, int group = 0) final;
-  int AddDeviceQuaternion(Vector3 position, Quaternion quaternion, int group = 0) final;
+  size_t AddDevice(Vector3 position, Vector3 euler_angles, size_t group = 0) final;
+  size_t AddDeviceQuaternion(Vector3 position, Quaternion quaternion, size_t group = 0) final;
 
-  const int numDevices() noexcept final;
-  const int numTransducers() noexcept final;
-  int GroupIDForDeviceIdx(int device_id) final;
-  const Vector3 position(int transducer_idx) final;
-  const Vector3 local_position(int device, Vector3 global_position) final;
-  const Vector3 direction(int transducer_id) final;
-  const Vector3 x_direction(int transducer_id) final;
-  const Vector3 y_direction(int transducer_id) final;
-  const Vector3 z_direction(int transducer_id) final;
-  const int deviceIdxForTransIdx(int transducer_idx) final;
+  const size_t numDevices() noexcept final;
+  const size_t numTransducers() noexcept final;
+  size_t GroupIDForDeviceIdx(size_t device_id) final;
+  const Vector3 position(size_t transducer_idx) final;
+  const Vector3 local_position(size_t device, Vector3 global_position) final;
+  const Vector3 direction(size_t transducer_id) final;
+  const Vector3 x_direction(size_t transducer_id) final;
+  const Vector3 y_direction(size_t transducer_id) final;
+  const Vector3 z_direction(size_t transducer_id) final;
+  const size_t deviceIdxForTransIdx(size_t transducer_idx) final;
 
  private:
   std::vector<std::unique_ptr<Device>> _devices;
-  std::map<int, int> _group_map;
+  std::map<size_t, size_t> _group_map;
 };
 
 GeometryPtr Geometry::Create() { return std::make_shared<AUTDGeometry>(); }
 
-int AUTDGeometry::AddDevice(Vector3 position, Vector3 euler_angles, int group) {
-  const int device_id = static_cast<int>(this->_devices.size());
+size_t AUTDGeometry::AddDevice(Vector3 position, Vector3 euler_angles, size_t group) {
+  const size_t device_id = this->_devices.size();
   const auto pos = Eigen::Vector3d(position.x(), position.y(), position.z());
   const auto ea = Eigen::Vector3d(euler_angles.x(), euler_angles.y(), euler_angles.z());
   this->_devices.push_back(Device::Create(pos, ea));
@@ -107,8 +107,8 @@ int AUTDGeometry::AddDevice(Vector3 position, Vector3 euler_angles, int group) {
   return device_id;
 }
 
-int AUTDGeometry::AddDeviceQuaternion(Vector3 position, Quaternion quaternion, int group) {
-  const auto device_id = static_cast<int>(this->_devices.size());
+size_t AUTDGeometry::AddDeviceQuaternion(Vector3 position, Quaternion quaternion, size_t group) {
+  const auto device_id = this->_devices.size();
   const auto pos = Eigen::Vector3d(position.x(), position.y(), position.z());
   const auto qua = Eigen::Quaterniond(quaternion.w(), quaternion.x(), quaternion.y(), quaternion.z());
   this->_devices.push_back(Device::Create(pos, qua));
@@ -116,19 +116,19 @@ int AUTDGeometry::AddDeviceQuaternion(Vector3 position, Quaternion quaternion, i
   return device_id;
 }
 
-const int AUTDGeometry::numDevices() noexcept { return static_cast<int>(this->_devices.size()); }
+const size_t AUTDGeometry::numDevices() noexcept { return this->_devices.size(); }
 
-const int AUTDGeometry::numTransducers() noexcept { return this->numDevices() * NUM_TRANS_IN_UNIT; }
+const size_t AUTDGeometry::numTransducers() noexcept { return this->numDevices() * NUM_TRANS_IN_UNIT; }
 
-int AUTDGeometry::GroupIDForDeviceIdx(int device_idx) { return this->_group_map[device_idx]; }
+size_t AUTDGeometry::GroupIDForDeviceIdx(size_t device_idx) { return this->_group_map[device_idx]; }
 
-const Vector3 AUTDGeometry::position(int transducer_id) {
+const Vector3 AUTDGeometry::position(size_t transducer_id) {
   const auto local_trans_id = transducer_id % NUM_TRANS_IN_UNIT;
   const auto pos = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->global_trans_positions.col(local_trans_id);
   return Vector3(pos.x(), pos.y(), pos.z());
 }
 
-const Vector3 AUTDGeometry::local_position(int device_idx, Vector3 global_position) {
+const Vector3 AUTDGeometry::local_position(size_t device_idx, Vector3 global_position) {
   const Device* device = this->_devices.at(device_idx).get();
   const auto local_origin = device->global_trans_positions.col(0);
   const Eigen::Vector3d x_dir = device->x_direction;
@@ -139,22 +139,22 @@ const Vector3 AUTDGeometry::local_position(int device_idx, Vector3 global_positi
   return Vector3(rv.dot(x_dir), rv.dot(y_dir), rv.dot(z_dir));
 }
 
-const Vector3 AUTDGeometry::direction(int transducer_id) { return z_direction(transducer_id); }
+const Vector3 AUTDGeometry::direction(size_t transducer_id) { return z_direction(transducer_id); }
 
-const Vector3 AUTDGeometry::x_direction(int transducer_id) {
+const Vector3 AUTDGeometry::x_direction(size_t transducer_id) {
   const Eigen::Vector3d dir = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->x_direction;
   return Vector3(dir.x(), dir.y(), dir.z());
 }
 
-const Vector3 AUTDGeometry::y_direction(int transducer_id) {
+const Vector3 AUTDGeometry::y_direction(size_t transducer_id) {
   const Eigen::Vector3d dir = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->y_direction;
   return Vector3(dir.x(), dir.y(), dir.z());
 }
 
-const Vector3 AUTDGeometry::z_direction(int transducer_id) {
+const Vector3 AUTDGeometry::z_direction(size_t transducer_id) {
   const Eigen::Vector3d dir = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->z_direction;
   return Vector3(dir.x(), dir.y(), dir.z());
 }
 
-const int AUTDGeometry::deviceIdxForTransIdx(int transducer_id) { return transducer_id / NUM_TRANS_IN_UNIT; }
+const size_t AUTDGeometry::deviceIdxForTransIdx(size_t transducer_id) { return transducer_id / NUM_TRANS_IN_UNIT; }
 }  // namespace autd
